requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements-dev.in
#
astor==0.8.1
    # via flake8-qgis
attrs==21.4.0
    # via
    #   flake8-bugbear
    #   pytest
black==22.1.0
    # via -r requirements-dev.in
cfgv==3.3.1
    # via pre-commit
click==8.0.4
    # via black
coverage[toml]==6.3.2
    # via pytest-cov
distlib==0.3.4
    # via virtualenv
filelock==3.6.0
    # via virtualenv
flake8==4.0.1
    # via
    #   -r requirements-dev.in
    #   flake8-annotations
    #   flake8-bugbear
    #   flake8-polyfill
    #   flake8-qgis
    #   pep8-naming
flake8-annotations==2.7.0
    # via -r requirements-dev.in
flake8-bugbear==22.1.11
    # via -r requirements-dev.in
flake8-polyfill==1.0.2
    # via pep8-naming
flake8-qgis==1.0.0
    # via -r requirements-dev.in
identify==2.4.12
    # via pre-commit
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements-dev.in
mccabe==0.6.1
    # via flake8
mypy==0.941
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
nodeenv==1.6.0
    # via pre-commit
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
pep8-naming==0.12.1
    # via -r requirements-dev.in
platformdirs==2.5.1
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.17.0
    # via -r requirements-dev.in
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pyparsing==3.0.7
    # via packaging
pytest==7.1.1
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-qgis
pytest-cov==3.0.0
    # via -r requirements-dev.in
pytest-qgis==1.3.1
    # via -r requirements-dev.in
pyyaml==6.0
    # via pre-commit
six==1.16.0
    # via virtualenv
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   mypy
    #   pytest
typing-extensions==4.1.1
    # via
    #   black
    #   mypy
virtualenv==20.13.3
    # via pre-commit
